The `AppShell.xaml` file has been modified to change the `Shell.FlyoutBehavior` property from "Flyout" to "Disabled", disabling the flyout menu in the application shell. Additional validation logic has been added to the `EditContactViewModel` class. Before editing a contact, the code now checks if any of the contact fields (`FirstName`, `LastName`, `Email`, `PhoneNumber`) are empty. If any field is empty, an alert is displayed to the user indicating that all fields must be filled in.
New validation checks for the email and phone number fields have also been added. If the email is not valid according to `ValidationHelper.IsValidEmail`, an alert is displayed to the user to enter a valid email address. Similarly, if the phone number is not valid according to `ValidationHelper.IsValidPhoneNumber`, an alert is displayed to the user to enter a valid phone number.